+++ src/test-resources/test_preproc_v3.h
@@ -1,0 +1,4 @@
+#pragma include_once
+void once_func() {
+    return;
+}